requirements.txt
allure-behave>=2.13.2
allure-combine==1.0.11
wsproto==1.2.0
behave>=1.2.6
selenium>=4.11.2
urllib3==1.26.18
requests==2.31.0
parse==1.19.1
